Refactor core drawing functionality into separate material (#299)
