+++ packages/ui/src/lib/components/Stage/components/FogOfWarLayer/FogOfWarLayer.svelte
@@ -6,10 +6,10 @@
   import type { Size } from '../../types';
   import type { Callbacks } from '../Stage/types';
   import LayerInput from '../LayerInput/LayerInput.svelte';
-  import FogOfWarMaterial from './FogOfWarMaterial.svelte';
   import toolOutlineVertexShader from '../../shaders/default.vert?raw';
   import toolOutlineFragmentShader from '../../shaders/ToolOutline.frag?raw';
   import { SceneLayer } from '../Scene/types';
+  import FogOfWarMaterial from './FogOfWarMaterial.svelte';
 
   interface Props extends ThrelteProps<typeof THREE.Mesh> {
     props: FogOfWarLayerProps;
@@ -53,7 +53,7 @@
     if (!isActive) {
       lastPos = null;
       drawing = false;
-      material?.render('revert', false);
+      material?.revertChanges();
     }
   });
 
@@ -102,7 +102,7 @@
     lastPos = null;
     drawing = false;
     outlineMaterial.visible = false;
-    material?.render('revert', false);
+    material?.revertChanges();
   }
 
   function draw(e: Event, p: THREE.Vector2 | null) {
@@ -133,7 +133,7 @@
    * Clears all fog, revealing the entire map underneath
    */
   export function clearFog() {
-    material?.render('clear', true);
+    material?.clear();
     onFogUpdate(toPng());
   }
 
@@ -141,7 +141,7 @@
    * Resets the fog to fill the entire layer
    */
   export function resetFog() {
-    material?.render('fill', true);
+    material?.fill();
     onFogUpdate(toPng());
   }
